io_queue: Do not use write bandwidth/iops as limit
To limit requests scheduler operates on two values:
- limit
- write-multiplier (just multiplier below)

The limit for both -- throughput and iops -- is currently taken
as the maximum of read and write measurements. The multiplier is
the "weight" of the write request, whil reads a weighed as 1.
When scheduling the sum of requests weights is maintained not to
go above the limit.

Let's take a disk with read iops 200 and write iops 100. The limit
is then set to 200 and the multiplier to 2 (read-iops / write-iops).
Respectively, the queue allows for 200 / 1 = 200 read requests and
200 / 2 = 100 writes.

This math breaks when the disks write speed is larger than the
read speed. Reversing the above example, let's take a disk with
read iops 100 and write iops 200. The limit would be still 200, but
the multiplier becomes 0.5. Respectively, the number of reads is
200 / 1 = 200 and the maximum number of writes is 200 / 0.5 = 400.
Both are 2x times larger than they should be.

The fix is to treat read speed as the limit and use write speed only
to get the multiplier.

Signed-off-by: Pavel Emelyanov <xemul@scylladb.com>