iostream: Use bool_classes to construct output_stream
The new constructor cannot have default arguments for its both
boolean options, otherwise it will conflict with the old one
being deprecated. To overcome this both of them have the first
arg not-default and the new temporary constructor without extra
arguments is added.

On option to make only the new constructor not have the default
first option won't work as it will deprecate also the case when
the object is constructed without bool options at all.

Signed-off-by: Pavel Emelyanov <xemul@scylladb.com>