iostream: Discover the buffer size from sink
On the output path the buffer size is only used by the stream
level. On the over hand it's the sink that knows the buffer
size, but it doesn't use it (data-source does). Also it's the
file-sink as well that has the file at hands and can adjust
the buffer size not to be larger than the maximum IO length.

In scylla output_streams often sits on top of a pile of sinks,
e.g. like

  output_stream -> compression_sink -> file_sink -> file

To properly configure the top-most output_stream one would
need to pick the lowest file (or, at least) the file_sink,
which can be quite complicated.

This propagation can be greatly simplified by introducing the
sink-only constructor of output_stream and the .buffer_size()
method for the sink. The file_sink will then get _its_ buffer
size and cap it with the file's maximum IO size, the stacked
sinks will forward the call downwards.

Signed-off-by: Pavel Emelyanov <xemul@scylladb.com>