io_queue: Move request weighting code into group
The logic that calculates the ticket for a given requests depends
purely on queue's config which now lives on the group, so it's
natural to make it a group's method.

Also, while at it, split the ticket generation method into two,
since its first half is going to complicate soon and the mixed
mode patches will need the 2nd half as well.

Signed-off-by: Pavel Emelyanov <xemul@scylladb.com>