io_queue: Keep ticket on queued request
The fair queue entry is going to stop carrying the ticket on-board, but
this value is needed e.g. to correctly cancel the request, so keep the
ticket on queued request entry.

This change allows the fair_queue_entry not to expose its ticket via
the public method.

A note on cancelling the queued request. When this happens the carried
request is kept as is though it's deaccounted from the fair queue and
its value is no longer needed. When the cancelled request is dispatched
(for removal) the fair-queue is notified with explicitly empty ticket
for the same reason.

Signed-off-by: Pavel Emelyanov <xemul@scylladb.com>