io_queue/fair_queue: Carry execution ticket down to dispatch
When executing a request will have another ticket value that differs
from the queued ticket one. This another ticket value needs to be
saved until request completion. At the same time this value cannot
be saved by the io_queue at the time it generates one, because
fair_queue may reject dispatching request with it and may come
later.

Said that, this patch prepares the gound for keeping the execution
tickets on the io-queue level by carrying it as an argument of
the io_queue dispatching calls.

Signed-off-by: Pavel Emelyanov <xemul@scylladb.com>