io_queue: Carry execution ticket down to dispatch
When executing the request needs to keep its dispatching ticket
on board to properly release the grabbed capacity when completed.
This value cannot be saved by the io_queue at the time it is
generated, because fair_queue may reject dispatching request and
may ask it again later.

Said that, this patch propagates the dispatching ticket from the
fair-queue code down to the io_desc_read_write.

Signed-off-by: Pavel Emelyanov <xemul@scylladb.com>