io: Introduce io_request::direction
It's very often needed to know whether the IO request is read OR write
and in most of these cases there's no 3rd option, despite it may exist,
e.g. sync one. This "3rd option" makes all the read-or-write checkers
exploit an ugly

    if (is_read() {
        ...
    } else if (is_write()) {
        ...
    } else {
        thrown some_exception();
    }

construct.

This patch introduces a boolean "direction" for the io_request. This
change simplifies the existing processing a bit and makes further
patching much simpler.

Signed-off-by: Pavel Emelyanov <xemul@scylladb.com>