io: Introduce direction-and-lengh
Next code will actively use the pair of request length and the
bit of whether the request is read or write. To make the future
life simpler introduce a helper class that carries both.

To save some space both values are put into a signed size_t
variable and the negative size means that the request is for
write.

Signed-off-by: Pavel Emelyanov <xemul@scylladb.com>