io_queue: Keep direction and length on desc
The io_desc_read_write represents an executing request. When it
completes it will notify the io_queue (and indirectly the fair
queue too) about it so that they de-account it as they want. This
completion notification needs both -- request direction and its
length. Both bits are known at desc construction time. This
change also simplifies the desc dispatch method a bit.

Signed-off-by: Pavel Emelyanov <xemul@scylladb.com>