resource: Move vector of queues into io_queue_topology
The io_queue_topology describes "numerical" topology -- the
numbers of queues and groups and their mappings to each other.
Later this info is materialized into device_io_topology by
creating vectors of queues and groups.

There's no real need in this split. The io_queue_topology can
just create the needed vectors with no objects in them, so
that later queues and groups are put into _it_ instead of on
the device_io_topology.

This is step #1 -- move vector of queues into io_queue_topology.

Signed-off-by: Pavel Emelyanov <xemul@scylladb.com>