resource: Configuration of mixed mode
All the examined disk models behave similarly in mixed mode handling.
Depending on the maximum write request size (a bit more on this below)
the bandwidth of the reads is reduced as compared to the pure read
bandwidth. The write bandwidth itself doesn't change, but works as if
it was competing with reads in a "fair" manner.

The mixed mode configuration is the list of

    maximum write length
    read bandwidth step-function

pairs. The steps can be different depending on the maximum size
of the write request in the disk. This looks counter intuitive,
but other options -- depend on the total number of writes in disk
OR on the total size of writes (or both) work worse. The numbers
that I see from disks show that the ratio of read bandwidth to
write bandwidth remains "expected" when chanhing the read or write
input flows, but the size of the write requests changes this ratio.

Signed-off-by: Pavel Emelyanov <xemul@scylladb.com>