io_queue: Mixed mode ticketing
That's the core of the mixed mode scheduling. It consists of 2 parts.

1. Mixer accounting

The mixer counter tracks whether there are writes at all in the
disk and what's the amount of reads in it. The former is needed
to check if the disk is in mixed mode at all and the latter is
to select proper read-request multiplier.

2. Execution ticket

Write requests use their "queued" ticket values as they are not
affected by reads anyhow special. Common proportions work for
them. Read requests get increased size weights according to the
mixed mode staircase bandwidth function.

Signed-off-by: Pavel Emelyanov <xemul@scylladb.com>