io_queue: Use queued ticket for shares acct
In mixed mode the io-queue will give larger dispatch tickets for
entries so that the capacity management code will put _less_
requests into the io-sink. However, it's not correct to use the
larger ticket to account for consumed resources, because in this
case the corresponding priority-class will look as if it consumed
more bandwidth than it really did. E.g. the dispatch ticket for
a read request may become up to 10 times larger. Respectively,
the class would be accounted for 10 times more accumulated points
and the fairness algo would give this class 10 times less disk
bandwidth.

Not to make this happen, teach the fair-queue code that the
dispatched requests have two tickets -- one is used for capacity
calculations and the other one for consumed shares.

Signed-off-by: Pavel Emelyanov <xemul@scylladb.com>