fair_queue: Account for changed ticket when pending
When attempting to dispatch a request the fair-queue code first
tries to account it into the shared capacity counters. For this
the dispatching ticket values is used. In case the ticket doesn't
fit into the capacity the queue is put into the pending state so
that it will try to push the request some time later (maybe several
times). When this "some time later" will come the request's
dispatching ticket value will probably change, so the capacity
checks will be performed with the value that may differ from what
the pending state descriptor has. When the request finally gets
through it will be dispatched and (!) completed with potentially
different value as well.

To keep the capacity accounting correct, when the request gets
dispatched the difference between pre-grabbed and dispatching
ticket should be correctly (de-)accounted into the group.

For now this code doesn't work, because requests are dispatched
with the same "constant" ticket values that they get when queued,
so it's also a preparatory patch.

Signed-off-by: Pavel Emelyanov <xemul@scylladb.com>