fair_queue: Shuffle dispatching loop
Right now its a loop in a loop that uses statistics to check if
it needs to continue. After the change it's a single loop and
statistics is _just_ statistics that doesn't take part in loop
decisions.

As a nice side effect this makes code shorter which is good for
the next patch that moves dispatch_requests into header.

Signed-off-by: Pavel Emelyanov <xemul@scylladb.com>