smp: Use after free in smp::invoke_on_others
The call sequence looks like this

smp::invoke_on_others(Func func)
 `- parallel_for_each([func = std::move(func)]
     `- smp::submit_to(func)

The func argument is first moved onto parallel_for_each's lambda then
the captured value is passed as an argument to the smp::submit_to. At
this point the func is considered as an lvalue reference on the value
that sits on the parallel_for_each's lambda capture. Respectively the
smp queue's async_work_item saves this reference onboard and the func
itself gets destroyed on exit from invoke_on_others.

Fix this by forcing copy of func when pushing it into submit_to. This
makes submit_to treat func as rvalue reference and save it (by moving)
on the async_work_item.

Since the fix makes invoke_on_others look like invoke_on_all in this
place, also tune the rest of it -- turn func argument into forwardable
reference and capture it by reference on the lambda. After it all the
invoke_on_... in smp toss the func the same way.

fixes: #946

Signed-off-by: Pavel Emelyanov <xemul@scylladb.com>