fair_queue: Add rate limiter
Based on a leaky bucket approach. Each priority class is equipped with
a limiter that carries a credit on board. It tries to consume from the
credit every time it wants to dispatch a requ^w entry. When the credit
is exhausted the priority class is "delayed" i.e. moved into a delayed
list. Classes (and rate limiters owned by them) are refilled from time
to time at the configured rate.  When the accumulated credit exceeds a
certain limit the accumulation stops.

All the aforementioned credit, limit and rate are fair queue tickets.
Next patch will introduce conversion from human readable bytes/sec and
ops/sec into tickets.

The "from time to time" refil can happen in two places. First is when
the prio calss is on the delayed list. This list gets refilled before
fair-queue is polled for dispatch. Second, the credit can be refilled
at the time of dispatch on not-yet-delayed class.

Also the io-queue enter to the interrup mode takes into account delayed
classes to wake up the reactor at the time they will have chance to get
refilled.

Signed-off-by: Pavel Emelyanov <xemul@scylladb.com>