io_queue: Account r/w separately
Now when the request direction is available during dispatch it's
possible to account for reads/writes bytes/ops separately.

This knowledge is very useful because, as we've learned, writes
inhibit reads in mixed IO resulting in increased read latencies.
Telling write throughput from read throughput in stats will help
to understand what's going on in the system better.

Signed-off-by: Pavel Emelyanov <xemul@scylladb.com>