fair_queue: Simplify base re-evaluation
Each time an accumulator get infinite the normalization lowers them
all and forwards the base time. This forwarding goal is to make it
as close to now as possible and a specially crafted math is used. For
different accumulator types this math may be not as precie, just
setting it to 'almost now' works in all cases.

Signed-off-by: Pavel Emelyanov <xemul@scylladb.com>