seastar: Introduce an option to force explicit IO class argument
Right now all IO methods from file and fstreams may silently perform
the IO inside default priority class. That's not a huge problem at
the moment, but creates a risk that some code starts doing badly
classified IO and it would be quite hard to spot it.

The new option keeps the default priority class on board, but removes
the default arguments from file/fstream methods so that callers will
have to explicitly (mis-)use the default IO class if it wants.

Signed-off-by: Pavel Emelyanov <xemul@scylladb.com>