fstream: Keep io_priority_class on internal classes
Now they keep it as a part of options, but the priority class is
going to become the argument to fstream creation functions and
thus disappear from all options. This patch prepares fstream core
code for that change by keeping the priority class on internal
data structures apart from options.

Signed-off-by: Pavel Emelyanov <xemul@scylladb.com>