code: Explicitly use default_priority_class() in core code
The IO methods from file/streams are going to optionally drop
the implicit default_priority_class() argument. Prepare for that
in some places.

HTTPD code may want to open files for reading. It would be nice
to have the ability to specify the class to use for that, but
for now default class would work.

The TLS code opens files with certs/trusts/keys/etc. Default IO
class works here as well.

Signed-off-by: Pavel Emelyanov <xemul@scylladb.com>