io_queue, fair_queue: Make fair_queue a template
For now the whole "templatezation" of the fair_queue works because of
the fair_queue_entry is sitting in between. The goal is to make fair
queue work directy on the entry type. To complete this change -- remove
the fair_queue_entry and patch all the fair_queue_impl code to use
the Entry type.

The queued_io_request is patched respectively to be the inheritant of
the slist_base_hook and provide the needed methods (most of them it
already provides).

Signed-off-by: Pavel Emelyanov <xemul@scylladb.com>