fair_queue: Remove statistics
It was used by the dispatch loop to decide whether or not to dispatch more,
but now it's purely statistical counters. The counters in question are
duplicated by the io_queue layer, so there's little sense in having both.
Another reason for removing it is that the upcoming io-queue improvement
blurs the concept of "fair_queue_ticket" and it makes even less sence in
keeping these stats.

Signed-off-by: Pavel Emelyanov <xemul@scylladb.com>