io_queue, fair_queue: Simplify cancellation
The notify_request_cancelled is needed to reset the ticket to zero. This
can be done directly on the fair_queue_entry. The notify_request_finished
was actually called to update "queued" statistics, it did nothing more
becase the passed ticket was zero. Now when the statistics is gone this
call is no-op.

Removing both helps next patches.

Signed-off-by: Pavel Emelyanov <xemul@scylladb.com>