fair_queue: Make maybe_ahead_of() non-method helper
The maybe_ahead_of calculates the difference between two wrapping
counters for the case when the counter is fair queue ticket. Soon
the counter will become uint64_t for which it will be impossible
to define a method. Turn the method into a standalone function, so
that it will be possible to overload it for uint64_t.

Rename it while at it.

Signed-off-by: Pavel Emelyanov <xemul@scylladb.com>