fair_queue: Rate-limiter based scheduling
** The math **

   Bw, Br -- write/read bandwidth (bytes per second)
   Ow, Or -- write/read iops (ops per second)

   BW, BR, OW, OR -- their maximum values (configured)

Formula:

   Bw/BW + Br/BR + Ow/IW + Or/IR <= K

K is the scalar value ~0.9

Next, bandwidth is bytes time derivatite, iops is ops time
derivative, i.e. Bx = d(bx)/dt, Ox = d(ox)/dt. Then the limiting
formula turns into

  d(bw/BW + br/BR + ow/OW + or/OR)/dt <= K

Fair queue tickets are {s, w} size-weight pairs that are

  s = read_base_count * br, for reads or
      BR/BW * read_base_count * bw, for writes

  w = read_base_count, for reads or
      OR/OW * read_base_count, for writes

So the formula turns into

  d(s/S + w/W)/dr <= K

where {S, W} is the {BR, OR} ticket (called TICKET below)

The dF/dt <= K limitation is managed by the leaky bucket algo with
K being the rate and F being the cumulative counter. To enforce our
formula we need to leaky-bucket the

   accumulate{ ticket.normalize(TICKET) }

value with the K factor.

The normalization results in a float of the 2^-20 order of magnitude.
Not to invent float point atomic arithmetics, the result is converted
to an integer by multiplying by a factor that's large enough to turn
2^-20 into an integer.

** Now the patch comment itself **

According to the math, this patch makes fair_group::capacity_t be
uint64_t and applies leaky-bucket to it.

The capacity is grabbed as usually -- per-request -- but is NOT released
upon request completion. Instead, each shard tries to replenish it
according to the last replenished timestamp.

Next, the leaky-bucket algo has 3 parameters:

- rate -- when replenished the credit is increased by time delta * this
  value. It's calculated as K * fixed_point_factor according to the
  math. The increase speed itself is controlled by the TICKET, which is
  now the fair group configuration parameter, calculated by reactor
  start up code from io-properties.yaml file

- limit -- the maximum value the credit may acheive. This corresponds to
  the maximum number of tickets that can be dispatched in one tick and
  it's inherited from the current model to meet the latency goal

- threshold -- the minimal number of accumulated cretids to allow for
  the credit refill. For now it's 1, but the goal is to make it larger
  to reduce the number of atomic updates on the replenish timestamp

Signed-off-by: Pavel Emelyanov <xemul@scylladb.com>