fair_queue: Introduce ticket capacity
The fair_queue ticket is used in two meanings in fair queue code

- a ticket to split the disk accurding to classes' shares
- a value to maintain disk throughput capacity

For now the throughput capacity is modelled with the ticket itself,
but further patching is turning it into a uint64_t. This patch
adds the convertion from raw ticket value into whatever capacity
is (now it just retrusn the ticket itself).

Signed-off-by: Pavel Emelyanov <xemul@scylladb.com>