fair_queue: Rate-limiter based scheduling
** The math **

   Bw, Br -- write/read bandwidth (bytes per second)
   Ow, Or -- write/read iops (ops per second)
   xx_max -- their maximum values (configured)

Throttling formula:

   Bw/Bw_max + Br/Br_max + Ow/Ow_max + Or/Or_max <= K

where K is the scalar value <= 1.0. It can be turned into the

  d(sum(w/W + s/S))/dr <= K

where {w, s} is the ticket value of a request and sum summarizes the
ticket values from all the requests seen so far, {W, S} is the ticket
value that corresonds to a virtual summary of Or_max requests of
Br_max size total.

** Now the patch comment itself **

According to the math, this patch makes fair_group::capacity_t be
uint64_t and applies token-bucket to it with 3 parameters

- rate -- when replenished the credit is increased by time delta * this
  value. It's K according to the math. The token value of each request
  is normalized by the _cost_capacity variable ({W, S} from the formula),
  which is calculated using the values taken from io-properties.yaml file

- limit -- the maximum value the credit may acheive. This corresponds to
  the maximum number of tokens that can be accumulated for the latency
  goal duration, i.e. it's latency_goal * rate

- threshold -- the minimal number of accumulated cretids to allow for
  the credit refill. For now it's 1, but the goal is to make it larger
  to reduce the number of atomic updates on the replenish timestamp

When refilling tokens two buckets are in use. Requests grab tokens from
the first bucket (if it has them) and return it back into the second.
The time-based replenisher moves tokens from the second bucket into the
first. This thing is needed to protect disk from the snow-ball effect.

When the number of requests in disk increase their latency goes up non
linearly. Spontaneous overloading a disk with few requests cause the
disk to slow down, time-based replenisher would generate more tokens
for that time and the disk would be overloaded even more.

Another thing to mention is that the above math implies float numbers
in use. E.g. typical token value is ~2^-30 seconds order of magnitude.
Not to invent float point atomic arithmetics, the floats are converted
to an integer by multiplying by a factor that's large enough to turn
these values into a non-zero integer.

And the last but not least, the bytes/ops are adjusted by io-queue by
multiplying them by a value ~128. When this is applied to the
normalization ticket the value becomes too large for 32-bit. Thus the
math is taken not in units/second, but units/rate_resolution where the
rate_resolution duration is selected to be small enough (milliseconds).

Signed-off-by: Pavel Emelyanov <xemul@scylladb.com>