io_queue: Calculate max ticket from fq
If a too large request is submitted into the fair-queue it can lock
because that single request would be too large for the maximally
available capacity.

Before the rate-limited scheduler the maximum capacity was known and
it was excactly the value from config. In the new scheduler the maximum
capacity is expressed in terms of fair_group tokens and io-queue
cannot "guess" it (without duplicating the code).

Said that, this change is in re-evaluating the io-queue-side ticket
limit with the fair-group help. Actually this is a temporary kludge,
io-queue should stop calculating fair-queue tickets and calculate
the fgroup capacity early.

Signed-off-by: Pavel Emelyanov <xemul@scylladb.com>