fair_queue: Adjust duration/factor for smaller rates
The fair queue math calculates the costs/rates with usec resolution.
This means that disks should have IOPS and bandwidth as large as to
produce non-zero costs each microsecond. This means at least 8k IOPS
(and 4MB/s) which can be a bit too much for cloud disks and HDDs.

Change the resolution to be milliseconds and adjust the float point
factor accordingly.

While at it -- make sure the cost capacity doesn't become zero for
slow disks.

Signed-off-by: Pavel Emelyanov <xemul@scylladb.com>