io_queue: Check latency goal and warn the user
If the disk is too slow it may result that even the smallest request
cannot get into the disk. In the previous scheduler it was checked by
the max-counts values, but today the more relevant check is the rate
limiter replenishment limit value.

This check is performed in io_group construction when it tries to find
the maximum length of the requests. Now it's asserted that the minimal
request is at least 1k, but the better behavior is to ask the user to
increase the latency goal.

Signed-off-by: Pavel Emelyanov <xemul@scylladb.com>