io_queue: Register metrics by io_queue
After all the previous patches it's now possible to make the
priority class only return back a vector of metric definitions
and make the io_queue label them and register. Next patches
will extend the vector of metrics with those from fair group(s).

As a placeholder for the metrics group the pclass field is
still used :\

Signed-off-by: Pavel Emelyanov <xemul@scylladb.com>