iotune: Take care of io-properties(-file) options
There are three options in iotune that look alike:

  --io-properties
  --io-properties-file
  --properties-file

The first two are actually not supposed to be in use, as iotune
needs to measure disk speeds with as little IO scheduler on its
was as possible.

So this patch makes sure the --io-... and --properties-file are
not in used together and warn the user about potential misuse
otherwise.

Signed-off-by: Pavel Emelyanov <xemul@scylladb.com>