iotune: Measure O_DSYNC write speeds
Some disks show much slower write bandwidth and IOPS when the file is
opened with O_DSYNC. The IO-scheduler is built with the asusmption that
it's not the case and doesn't treat O_DSYNC IO anyhow special.

Teach iotune to measure the disk with O_DSYNC writes and at least warn
the user if the resulting performance deviates too much from the non
O_DSYNC one.

Signed-off-by: Pavel Emelyanov <xemul@scylladb.com>