io_queue: Don't let preemption overlap requests
When a queue tries to dispatch a request it first grabs the capacity
from group. When grabbing fails queue enters the pending state which
means that it queues itself at some position in the group. Next, the
group polls the group position for "my turn". When it happens the
queue may dispatch a request.

However, it's possible that on entering the pending state and on
exiting from it different requests with different capacities are in
use. Dispatching the new request "as is" is wrong -- upon completion
the new request would release some capacity that's different from
what was grabbed. If the released capacity is less than grabbed, the
queue can get locked as other shard(s) may wait for the larger shift
of the shared position. If the released capacity is greater than
grabbed, this will put extra capacity into the group. Accumulated
this would break all the configured dispatching limits.

There's a simple solution to this issue -- release the grabbed capacity
and try obtaining the new position for the new request. This is not
nice, as it will affect the latency of preempting requests the bad way
-- since entering the pending state other shards had likely queued
their requests and going to the tail of this queue means throwing
away the pending-wait time and accumulating the pending latency. Said
that -- current code tries to dispatch the preempted (new) request
right at once and somehow maintain the capacity difference.

Before rate-limiter approach both capacity and position in the shared
queue were very hard to estimate, because they were measured in 2d
tickets that don't have unambiguous less/greater/difference operators.
Not to mess with it the current code actually queued the old request
right after the new one. This trick guaranteed that the potential gap
between the other shard's queue position is covered, but also made
old request overlap its position with the position of requests from
other shards. Another problem with that code is that preemption could
happen several times in on go, thus potentially generating a long
queue of re-queued local requests and all this queue was overlapped
with other shards.

That was the intro :) Now the fix.

With the rate-limiter merged the capacity and position in the shared
queue are both 64-bit scalar values that can be compared. With this
ability the preemption handling is like this.

If the new request is smaller, it gets dispatched right at once, the
extra capacity is immediatelly released, the old request is re-queued
at the tail. This would work as if there were no attempt to dispatch
the old request, it would increase the old request latency, but since
it's larger it's likely the throughput-oriented request and doesn't
want low latency.

If the new request is larger, then it shouldn't be dispatched right
here, because it would overlap with another shard. In this case
queue should still try to dispatch the old request (which's smaller
and is thus likely small latency-sencitive read). For this the new
priority class should be temporarily withdrawn from the dispatching
queue.

The rl-iosched test results

* Over real disk (dev desktop)

** Pure read
                      iops    latency(us)       K
  read_rated:       219105      124.3       0.753

** Mixed, was:
                      iops    latency(us)       K
  write:             16506    28691.1       0.637
  read_rated:        96728     4664.3       0.332

** Mixed, patched:
                      iops    latency(us)       K
  write:              2080   147334.9       0.080
  read_rated:       161271     1788.8       0.554

* Over /dev/null with the same io-properties.yaml

** Pure read
                      iops    latency(us)       K
  read_rated:       219214        6.0       0.753

** Mixed, was:
                      iops    latency(us)       K
  write:             12329    51000.0       0.476
  read_rated:       148576     2019.1       0.511

** Mixed, patched:
                      iops    latency(us)       K
  write:              2037   157711.8       0.079
  read_rated:       169665     1504.2       0.583

(read class shares = 2500, write class shares = 100)

Signed-off-by: Pavel Emelyanov <xemul@scylladb.com>