io_queue: Pending needs to keep capacity instead of ticket
Next patch will need to compare pending ticket with the ticket of
the dispatching request for less/greater, which is impossible for
tickets, but it possible for capacities. So make this switch in
advance.

At this point this changes almost nothing, as different tickets
typically have different capacities. Of course it may happen that
capacities of different tickets coincide, but this is rare and
even if so, exiting from pending state with a different request
is perfectly safe.

Signed-off-by: Pavel Emelyanov <xemul@scylladb.com>