fair_queue: Dispatch request, _then_ re-queue the priority class
At the end of dispatch loop body the picked request is dispatched,
but before it its priority class is put back into the dispatch queue
if it still has any requests. This exact sequence is not important
now, but next patches will mark classes as "unplugged" inside the
dispatching callback and that should prevent the class from being
put into the dispatching queue. Swapping dispatch and re-queue makes
this work.

Signed-off-by: Pavel Emelyanov <xemul@scylladb.com>