fair_queue: Add class plug/unplug facilities
Nowadays the fair-queue priority class can be in one of two states --
either queued or not -- and it's if the class in question participates
in dispatching loop. Next patches will need one more state -- if the
class is allowed to be queued or not. This patch calls this (un)plugged.

The unplugged class can keep requests on board, but it doesn't get
involved into the dispatch loop until it's plugged back.

Signed-off-by: Pavel Emelyanov <xemul@scylladb.com>