io_queue: Bandwidth limiter
The limiter re-uses shared token bucket to rate-limit the priority
class. The token bucket is configured with a fixed rate, limit and
threshold, requests calculate the needed tokens like

    length / max_bandiwdth * fixed_point_factor;

This math is used in RL-scheduler already with slightly different
token calculation. Some differences still exist.

First, the token bucket is one-bucket and auto-releases its tokens.
This is done because there's no "consumer works slower than we
expect" problem here. Even if disk slows down the node-wide limiter
will handle it.

Next, the rate is updateable via API call. This is done by changing
the denominator of the above math, so the rate-limiter parameters
remain the same, but requests charge different costs from it.

Finally, the throttling is done is a "lazy" manner. Each time a
request is dispatched it accounts itself in the token bucket and if
the tokens deficiency appears on it, the request continues dispatch
but the class in question is unplugged from disk, so next requests
will come after tokens replenish. Unlike RL-sched-side tokens, here
they are replenished by timer.

Signed-off-by: Pavel Emelyanov <xemul@scylladb.com>